mixins fixed:
- two separate mixins.js created: dialogsMixins, pagesMixins;
- most methods moved to mixins;
- mixin methods renamed - 'mixin' removed
