+++ components/UsersDetailsDialog.vue
@@ -18,9 +18,9 @@
 </template>
 
 <script>
-import { Mixins } from "~/mixins/mixins";
+import { dialogsMixins } from "~/mixins/dialogsMixins";
 export default {
-  mixins: [Mixins],
+  mixins: [dialogsMixins],
   data() {
     return {
       entity: "users",
@@ -46,7 +46,7 @@ export default {
     async open(id) {
       if (id <= 10) {
         // Items with id < 10 are actual items, that exist on the server.
-        await this.mixinFindOne(id);
+        await this.findOne(id);
       } else {
         // Items with id > 100 don't actually exist on the server.
         // Clear the object properties > Display v-card with an error.
